Changes based on review feedback
